requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.1
flask-sqlalchemy==2.5.1
python-dotenv==0.19.0
bcrypt==3.2.0
PyJWT==2.3.0
